requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.1.0
tqdm==4.66.1
PyYAML==6.0.1
scipy==1.11.4
opencv-python==4.8.1.78
setuptools==65.5.0
gfpgan==1.3.8
realesrgan==0.3.0
gdown==5.1.0